test(bookmarks): split bookmark coverage by package
Why:
- App-host tests still carried bookmark persistence and reader bridge package logic, keeping this migration slice on the slowest simulator/app-host path.

What Changed:
- Moved BookmarkService persistence coverage into BibleCoreTests.
- Moved bookmark reader bridge, StudyPad, label, and accessibility coverage into BibleUITests and removed the old app-host source from the Xcode project.
- Kept JSword range coverage SWORD-backed and documented the remaining fixture coupling in the migration plan.

Validation:
- git diff --cached --check
- python3 scripts/check_repo_standards.py docblocks --all-files
- swift package describe --type json
- plutil -lint AndBible.xcodeproj/project.pbxproj
- xcodebuild test -scheme BibleCoreTests -only-testing:BibleCoreTests/BookmarkServicePersistenceTests
- xcodebuild test -scheme BibleUITests -only-testing:BibleUITests/BookmarkReaderBridgeTests
- xcodebuild build-for-testing -scheme AndBible
- npx gitnexus analyze
- GitNexus staged detect_changes reported LOW risk with no affected flows.

Impact:
- Reduces the app-host test bundle by one large bookmark file while preserving Android bridge and persistence contracts in package lanes.
- No production behavior change.
